+++ hippodamus-tests/test/dd/kms/hippodamus/testUtils/execution/configuration/TestAggregationConfigurationBuilder.java
@@ -20,8 +20,11 @@ public class TestAggregationConfigurationBuilder<S, R> extends BaseTestConfigura
 	@Override
 	public <T extends Throwable> ResultHandle<S> aggregate(ExceptionalCallable<S, T> callable) throws T {
 		TestCallable<S, T> testCallable = new TestCallable<>(testCoordinator, callable);
-		ResultHandle<S> handle = wrappedBuilder.aggregate(testCallable);
-		testCallable.setHandle(handle);
-		return handle;
+		return wrappedBuilder
+			.onHandleCreation(handle -> {
+				testCallable.setHandle(handle);
+				handleConsumer.accept(handle);
+			})
+			.aggregate(testCallable);
 	}
 }